edit

+++ frontend/src/components/SubcategoryPage.tsx
@@ -2,6 +2,8 @@ import React, { useEffect, useState } from 'react';
 import { useParams, Link, useLocation } from 'react-router-dom';
 import "./SubcategoryPage.css";
 import { useNavigate } from "react-router-dom"
+import Modal from 'react-modal'
+import CreateQuestion from './CreateQuestion';
 
 
 // types.ts
@@ -24,7 +26,7 @@ const SubcategoryPage: React.FC = () => {
     const navigate = useNavigate();
 
     const location = useLocation()
-    const subcategoryId = subcategory_id ? parseInt(subcategory_id, 10) : undefined;
+    const subcategoryId = subcategory_id ? parseInt(subcategory_id, 10) : 0;
     const categoryId = location.state as number;
     const ids = {
         subcategory_id: subcategoryId,
@@ -33,14 +35,19 @@ const SubcategoryPage: React.FC = () => {
 
     const [subCategoryName, setSubCategoryName] = useState<string>('');
     const [questionList, setQuestionList] = useState<Question[]>([]);
+    const [modalIsOpen, setModalIsOpen] = useState<boolean>(false);
 
-    // const handleQuestionClick = (question_id: number) => {
-    //     navigate(`/subcategory/${subcategory_id}`, { state: category.id });
-    // };
 
     // 削除ボタンを押すと、そのサブカテゴリーを削除する。
-    // その際、「削除してももんだいないか」確認メッセージを表示する。
+    // その際、「削除」と入力してクリックすることで削除が実行される。
     const handleDelete = async () => {
+
+        let confirmation = prompt("削除を実行するには、「削除」と入力してください:");
+
+        if (confirmation !== '削除') {
+            return;
+        }
+
         const response = await fetch(`http://localhost:8000/subcategories/${subcategory_id}`, {
             method: 'DELETE',
         });
@@ -73,25 +80,34 @@ const SubcategoryPage: React.FC = () => {
     }, [subcategory_id]);
 
     return (
-        <div>
+        <div className='subcategory-page'>
             <div className='subcategory-box'>
                 <h1>{subCategoryName}</h1>
-                <button onClick={handleDelete}>Delete</button>
+                <button className='delete-btn' onClick={handleDelete}>Delete</button>
             </div>
-            <Link 
+            {/* <Link 
                 to={{ pathname: "/createquestion" }}
                 state={ids}
             >
                 Questionを作成する
-            </Link>
-            {questionList.map((question) => (
-                <div className="questionBox" key={question.id}>
-                    <h2>{question.problem}</h2>
-                    {question.answer.map((answer, index) => (
-                        <p key={index}>{answer}</p>
-                    ))}
-                </div>
-            ))}
+            </Link> */}
+            <button onClick={() => setModalIsOpen(true)}>Questionを作成する</button>
+            <Modal
+                isOpen={modalIsOpen}
+                contentLabel="Example Modal"
+            >
+                <CreateQuestion category_id={categoryId} subcategory_id={subcategoryId} setModalIsOpen={setModalIsOpen}></CreateQuestion>
+            </Modal>
+            <div className='question-container'>
+                {questionList.map((question) => (
+                    <div className="question-box" key={question.id}>
+                        <h2>{question.problem}</h2>
+                        {question.answer.map((answer, index) => (
+                            <p key={index}>{answer}</p>
+                        ))}
+                    </div>
+                ))}
+            </div>
         </div>
     );
 };
